+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Enhancements
 
-* **Default overwrite to true**
+* **Remove `overwrite` field** from fsspec and databricks connectors
 * **Added migration for GitLab Source V2**
 
 ## 0.2.1